community_minutes/requirements.txt
Django>=1.9,<1.10
psycopg2>=2.6
django-compressor>=1.6
django-libsass>=0.6
django-debug-toolbar>=1.4
gunicorn>=19.1.1
Fabric>=1.10.0
pyocr==0.4.1
wand==0.4.3
wget==3.2
pdfminer3k==1.3.0
beautifulsoup4==4.5.1
requests==2.11.1
elasticsearch==2.4.0
